Retrabalho do front-end, retrabalho do back-end, refatoração do CSS e remoção de bugs

+++ front_end/src/components/EntradaDados/index.tsx
@@ -9,7 +9,7 @@ export const EntradaDados = (props: EntradaDadosProps) => {
   const [numeroDeEntrada, setNumeroDeEntrada] = useState(0);
 
   const changeNumeroHandler = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setNumeroDeEntrada(parseInt(event.target.value));
+    setNumeroDeEntrada(parseFloat(event.target.value));
   };
 
   const executarComEnter = (keyboardEvent: React.KeyboardEvent<HTMLInputElement>) => {
